Create a vendor chunk
This will allow us to load the entire brace library to get all the
themes and keyboard handlers without overly bloating the app chunk.

This is complicated because the vendor chunk includes the manifest by
default. Since the manifest changes every time the application code
changes, that's not very useful! Large thanks go to [SurviveJS] for
pointers towards the proper setup.

[SurviveJS]: http://survivejs.com/webpack/building-with-webpack/splitting-bundles/