Make `fork-bomb-prevention` a default feature
That way it is possible to not use the associateed
`--pids-limit` flag by building like so:

`cargo build --no-default-features`

This is useful for local sandbox installations running on a different
OS which doesn't directly support limiting PIDs, or which works
differently in that regards.

I am using coreOS for example, but admittedly also didn't try
to find a way to make it work.