Introduce more powerful routing
Routing with Redux appears to be the Wild West at the
moment. React-Router v4 is oh-so-close to completion, but the main
repository doesn't seem to want to talk about Redux, and has even
removed what was public because it was so out of date.

*Conceptually*, routing should be simple... "all we need to do" is
mirror our state in the URL and vice versa. Of course, there are
complications.

We take the approach of monitoring the Redux store. A user-provided
selector function extracts the interesting aspects of the state. When
that changes, we run another user-provided function to convert the
state into a Location object. This Location is pushed to the browser,
synchronizing the state to the URL.

We also monitor the browser's URL. When this changes, we run a
user-provided function that returns an optional action to
dispatch. This synchronizes the URL to the state.

We also provide a `Link` component that *duplicates* the current
state and runs an action. We then unconditionally run the
user-provided state-to-URL function, generating the appropriate HREF.