Add WebAssembly support to the docker containers
We add a custom `cargo wasm` command that behaves similar enough to
the existing `cargo rustc` to be used without changing too much
infrastructure. For right now, this is only supported in the nightly channel.

We also remove all crates instead of trying to figure out which ones
are valid. This leaves an ugly UI gap as we don't have a way of
supporting two distinct sets of crates at the moment.