Restore scrolling of the help page
We are using React-Portal to add the Popper.js Poppers to the DOM. By
default, these are added at the end of the `<body>` tag. Since our
content takes up the entire page, adding them causes the scrollbars to
appear momentarily, which Popper.js uses in the calculations. When the
popper is moved to the right place, the scrollbar disappears, and the
popper has to move a second time.

We "fixed" this by disabling the scrollbars completely, forgetting
about the help page. This fixes it by positioning the popper so it
doesn't cause the scrollbars to appear to start with.

Closes #304