Update top-crates to Cargo 0.29
Using cargo from git because the developers are in the process of
figuring out how to align their version numbers with rustc and may need
to skip the next one or two releases. The Manifest::custom_metadata
method is new in master and will be required for supporting
playground-specific custom metadata.