Fix failing tests
This commit fixes errors on a number of sandbox tests. An edition
was added to all compile and execute request tests, stable channel
received a value of None, the nighly test uses Edition::Rust2018.
In addtion, Clippy has changed the 'eq_op' lint to deny. Updated
the 'linting_code' test to expect this.