Debounce back-to-back change events from Ace
When we hit "return" when the cursor is indented, Ace generates two events:

1. The newline
2. The indentation

These events flow through Redux and are both applied to the editor as
they might have come from any other source. However, by the time we
apply them, Ace's internal state is up-to-date, so it looks like we
are reverting the changes, causing all sorts of strange behavior.

Back-to-back events can also come from "undo" / "redo" commands or
multi-cursor editing.

As a pretty good bandaid, this commit batches up the events before a
RequestAnimationFrame and only submits the last one. This isn't
*perfect* as mashing on the keyboard can still generate some issues.

Fixes #496