Prevent expanding the window size when adjusting the split
Since we were rotating the *entire* gutter, it would easily extend
beyond the window boundaries, adding a scrollbar to the entire
playground. Instead, rotate only the handle icon itself. Due to [a
limitation with split.js][limitation], we have to also ignore pointer
events on the handle since it's now a child of the gutter.

[limitation]: https://github.com/nathancahill/split/issues/354