Preparing for React 18: change how router decides to PUSH new URL
With React 18, the old hack of setting a boolean during the
browser-based update doesn't work anymore (for some reason)and a state
update comes in after we think the browser is done. This causes the
history stack to be destroyed so we can't go to the help page, back,
then forward again.

We now check to see if the Redux-driven URL would result in us ending
back at the same URL the browser does, and if so we just ignore the
update.