Support CompileRequest in the orchestrator
The orchestrator allows starting up a Docker container and then
communicating with it while it's running. We can then request files to
be read or written in the container as well as running processes.

Just enough has been done to connect the pieces needed to support
compiling to assembly / LLVM / MIR, etc. Future work will add other
operations including the ability to run the code.

Co-authored-by: Jake Goulding <jake.goulding@integer32.com>