Try to send larger stdout / stderr packets
Line-based messages were fine for small examples, but if someone
writes 10K lines, that translates to 10K WebSocket messages, which is
a bit excessive.

This tries to grab up to 32K of UTF-8 text at once, managing the
possibility of UTF-8 characters being split across the buffer.

Running the code `for _ in 0..128_000 { println!("a"); }` resulted in

Before: 71445 messages (and Chrome died)
After:   2631 messages