Don't build container tools atop playground sources
Originally, I did this to prevent downloading the crates multiple
times, which was useful. However, it means that the tools have to
rebuild when the crate list does, which is less helpful.

Moving the final `COPY`s later in the stage also avoids rebuilding the
dependencies when the tools change.