Encourage the OOM killer to pick the container processes
play.rust-lang.org has had a few outages because the machine is a bit
low on memory and the OOM killer picks to reap nginx. There are many
containers running using a small amount of memory while the one nginx
uses more, so it's a more attractive target.

My ideal situation would be if we could group all the containers and
kill from that pool until the memory pressure is resolved, but I don't
know how to do that.