Kill the container ourselves when done
We stream messages from the coordinator to the worker via stdin/stdout
piped to `docker run`. However, sometimes we close the pipe
corresponding to `docker run`'s stdout before it has had a chance to
fully write everything. If that happens, `docker run` exits without
deleting the container [1].

This has likely always been a problem with the coordinator / worker
architecture, but it's exacerbated by the recent addition of the
memory / CPU status messages that occur every second, which greatly
increase the amount of data crossing the boundary.

This reorganizes the code to pass in a known reasonably unique name to
the container and then attempts to kill that container at cleanup
time.

I tested by:

- Setting the worker status interval to 16ms (from 1s)
- Setting the webserver idle interval to 7s (from 60s)
- Setting the webserver session timeout to 14s (from 45m)
- Running code that prints infinitely (`loop { println!("moo"); }`)

This reliably recreated the `write /dev/stdout: broken pipe` error
from the `docker run` invocation, but after this commit there were no
vestigial running containers.

[1]: https://github.com/moby/moby/issues/38769