Prevent massive amounts of outgoing stderr / stdout
We've noticed some bad resource usage patterns in production which
manifests as 100% CPU usage, split evenly between system and user.

My hypothesis is that people have written some infinite loop that
prints output. We are able to transfer that data from the process to the
worker to the coordinator/ui to nginx to the user. However, the user
can't actually consume 10+ MiB of output.

There's also a possibility that Docker has effectively an infinite
buffer for stdout/stderr. In that case, we might just be bloating up
memory somewhere.

This is intended as a first step, to find people that are legitimately
trying to produce a bunch of output. From there, we could potentially
tweak the code to allow people to opt-in to generating even more data.