Address flaky execute_kill test
The primary loop in `spawn_cargo_task` works by getting messages from
the worker and then handing them off to the stdout / stderr / status
channels. If one channel is full, then everything will queue behind
it, resulting in all work stopping.

The test was waiting on the stdout channel and had dropped the status
channel, but nothing was processing the stderr channel. This means
that if enough[^1] stderr packets had been delivered before any
stdout, the test would fail.

To reproduce this, you can hack the loop in `spawn_cargo_task` so that
it delivers every `stderr` packet multiple times, forcing the channel
to be filled immediately.

[^1]: Over eight with the current setup.