feat: tari crates (#6)
* feat: add tari crates to top-crates

Adds a list of additional crates to install in the docker image for the
rust compiler.

* feat: add tari dependencies to docker

To run docker, we need some additional packages to be installed in the
Linux OS, incl protobuf, and sqlite.

Customise the security notice for Tari.

Add a README for the compiler

* fix: modify gha workflows for tari context

Editied workflows directly. CI will build everything when tagged with an
image or if the 'build' label is applied (nice for testing).

Don't run `ci/generate` since this will overwrite changes. I haven't
updated that utility yet.

* fix: add working tari crates

Added tari crates that are able to be published / are up to date
The rest are not yet on crates.io or are too old to be used.

* fix: target branch for prs should be master