Provide unique ids across `Coordinator` instances
Previously we tried to do this from inside the `Coordinator` /
`CoordinatorFactory` types, but that didn't handle the case when two
`CoordinatorFactory` instances were created (almost) simultaneously.

Instead, push the job of determining a unique ID outside of these
types, allowing it to be somewhere higher in the call stack (ideally
nigh-global).