Try to more thoroughly kill child Docker processes
During a recent outage, the Docker service reported 161 running
`docker` processes, while the playground was only aware of 27. Each
`docker` processes (and their children) use about 30 MiB of RAM. That
would total 4.7 GiB of memory, while the instance only has 4GiB.

My hypothesis is that we are starting containers, then encountering an
error and bubbling it up to the caller. Along the way, we skip the
call to shutdown the container.

This change tries to shutdown the container when the value is
dropped. This should allow it to be handled in error cases as well.

It's possible that this might also help with some of the zombie
processes we see from time-to-time!