Do not include optional crates that are not selected features
nalgebra released version 0.34 with a bunch of optional dependencies
that are not enabled by default[1]. The previous code didn't account
for non-selected optional dependencies, so it was pulling in all of
those versions.

This rework was required in order to get access to Cargo's "feature
resolver", which is the part we need to query to see if a dependency
is active or not.

Unfortunately, this required that we have a workspace and thus
physical files on disk, which means that the overall implementation is
much less elegant.

[1]: https://crates.io/crates/nalgebra/0.34.0/dependencies